chore: added script to verify commits
