+++ scripts/vcs/constants.ts
@@ -1,0 +1,1 @@
+export const PERMITTED_COMMIT_TYPES = ["feat", "chore", "fix", "docs", "test"];
